p2p/discover: Reorder symbols in udp.go
